requirements.txt
aiogram==3.0.0b7
sqlalchemy==2.0.25
pandas==2.1.4
python-dotenv==1.0.0
aiohttp==3.9.1